# samples/request-logger/src/openapi.yaml
openapi: 3.1.0
info:
  title: Book Request Logger
  version: 1.0.0
paths:
  /books:
    get:
      summary: List Books
      operationId: list_books_books_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      summary: Create Book
      operationId: create_book_books_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookPayload'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}:
    put:
      summary: Update Book
      operationId: update_book_books__book_id__put
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Book
      operationId: delete_book_books__book_id__delete
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BookPayload:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        year:
          anyOf:
          - type: integer
          - type: 'null'
          title: Year
      type: object
      title: BookPayload
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError